separated the function into different files
