+++ index.js
@@ -19,12 +19,12 @@ mongoose.connect(process.env.DB_CONNECT, {
 .catch(err => console.error('Failed to connect to MongoDB', err));
 
 // Import job-related functions
-const { addJob, viewJobs, editJob, deleteJob, getJobById } = require('./utils/JobUtil');
-app.post('/add-job', addJob);
-app.get('/view-jobs', viewJobs);
+const { editJob, getJobById } = require('./utils/update-job');
 app.put('/edit-job/:id', editJob);
 app.get('/view-job/:id', getJobById);
-app.delete('/delete-job/:id', deleteJob);
+
+const { viewJobs } = require('./utils/view-job');
+app.get('/view-jobs', viewJobs);
 
 app.get('/', (req, res) => {
     res.sendFile(__dirname + "/public/index.html");
